Refactor of Contact part on Profile page
